06-content-creation-agent/requirements.txt
crewai>=0.80.0
crewai-tools>=0.17.0
openai>=1.50.0
python-dotenv>=1.0.0
langchain-openai>=0.2.0
